pyproject.toml
[tool.poetry]
name = "matte"
version = "0.1.0"
description = ""
authors = ["Anuj Arora <anujarora920804@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^8.2.0"
numpy = "^1.20.3"
torch = "1.7.0"
torchvision = "0.8.1"

[tool.poetry.dev-dependencies]
black = "^21.5b0"
pylint = "^2.8.2"
isort = "^5.8.0"
pylint-fail-under = "^0.3.0"
coverage-badge = "^1.0.1"
pytest-cov = "^2.11.1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
